# Blocks: Short-circuit validation for identical markup